Se refactoriza la jerarquia de directorios para poder aplicar filtros de busqueda genericos
