--- app/ClinicSearch/Filters/BusinessName.php
@@ -1,1 +1,0 @@
-<?phpnamespace App\ClinicSearch\Filters;use Illuminate\Database\Eloquent\Builder;class BusinessName implements Filter{    /**     * Apply a given search value to the builder instance.     *     * @param Builder $builder     * @param mixed $value     * @return Builder $builder     */    public static function apply(Builder $builder, $value)    {		return $builder->where( 'business_name', 'like', '%' . $value . '%' );    }}
