+++ app/Searchers/ClinicSearch/Filters/BusinessName.php
@@ -1,0 +1,1 @@
+<?phpnamespace App\Searchers\ClinicSearch\Filters;use Illuminate\Database\Eloquent\Builder;use App\Searchers\Filter as Filter;class BusinessName implements Filter{    /**     * Apply a given search value to the builder instance.     *     * @param Builder $builder     * @param mixed $value     * @return Builder $builder     */    public static function apply(Builder $builder, $value)    {		return $builder->where( 'business_name', 'like', '%' . $value . '%' );    }}
